douban movice

+++ gankio/src/main/java/com/sky/gank/data/douban/DoubanMovieData.java
@@ -179,7 +179,7 @@ public class DoubanMovieData extends BaseResponse{
              */
 
             private int max;
-            private double average;
+            private float average;
             private String stars;
             private int min;
 
@@ -191,11 +191,11 @@ public class DoubanMovieData extends BaseResponse{
                 this.max = max;
             }
 
-            public double getAverage() {
+            public float getAverage() {
                 return average;
             }
 
-            public void setAverage(double average) {
+            public void setAverage(float average) {
                 this.average = average;
             }
